boot: Imaging validation progress
Use a `&'a RefCell<F>` for the flash device itself.  This adds a little
complexity, but we need to be able to share the mutability of the flash device
across different parts of the code, and this keeps it reasonably clean.

Implement an iterator for the TLV, and start the validate method.  This checks
that the SHA256 TLV entry is present, and is the only entry present.  The TLV
does not handle the protected TLVs yet.