storage: Create initial version of storage
`storage` attempts to be a version of embedded-storage::NorFlash and friends,
but with a few changes:

  - Errors aren't dynamic.  There is just a fixed set of errors.
  - The alignments are methods not associated constants.
  - usize instead of u32.

The last is just a matter of convenience.  Otherwise, this should allow for
`&dyn Flash`, which doesn't work with the embedded-storage types.  This is
needed for environments where the particular flash part is dynamically queried,
or for testing, where the code may want to test a large number of
configurations.

It should also save a lot in code space when using multiple flash partitions or
devices that have different parameters.